% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_functions.R
\name{create_OpenTree_df_decision}
\alias{create_OpenTree_df_decision}
\title{create_OpenTree_df_decision
\code{create_OpenTree_df_decision} post-processes the decision tree R dataframe}
\usage{
create_OpenTree_df_decision(df_input, df1)
}
\arguments{
\item{df_input}{input dataframe}

\item{df1}{side dataframe}
}
\value{
processed dataframe
}
\description{
create_OpenTree_df_decision
\code{create_OpenTree_df_decision} post-processes the decision tree R dataframe
}
