% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{create_tree}
\alias{create_tree}
\title{create_tree
\code{create_tree}}
\usage{
create_tree(file_name, dir_name)
}
\arguments{
\item{file_name}{name of the file}

\item{dir_name}{directory to save the file to}
}
\value{
opens OpenTree
}
\description{
create_tree
\code{create_tree}
}
