% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tree.R
\name{open_tree}
\alias{open_tree}
\title{open_tree
\code{open_tree}}
\usage{
open_tree(file_name, dir_name)
}
\arguments{
\item{file_name}{name of the file}

\item{dir_name}{directory to save the file to}
}
\value{
opens OpenTree
}
\description{
open_tree
\code{open_tree}
}
