% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_functions.R
\name{readOpenTreeModel}
\alias{readOpenTreeModel}
\title{readOpenTreeModel
\code{readOpenTreeModel} reads and converts JSON file into R dataframe}
\usage{
readOpenTreeModel(treeName)
}
\arguments{
\item{treeName}{name of the tree model}
}
\value{
a dataframe to be processed
}
\description{
readOpenTreeModel
\code{readOpenTreeModel} reads and converts JSON file into R dataframe
}
