% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTree.R
\name{runTree}
\alias{runTree}
\title{Vectorized categorical distribution}
\usage{
runTree()
}
\arguments{
\item{m.Probs}{matrix with probabilities (n.i * n.s)}
}
\value{
n.i x 1 matrix filled with sampled health state(s) per individual
}
\description{
\code{samplev} sample states for multiple individuals simultaneously.
}
