% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_functions.R
\name{evaluate_model}
\alias{evaluate_model}
\title{evaluate_model
\code{evaluate_model} evaluates the decision model}
\usage{
evaluate_model(treeName, params = list(), n_payoffs, backend = FALSE)
}
\arguments{
\item{params}{list of model input parameters}

\item{n_payoffs}{number of payoffs}

\item{backend}{whether to return the backbone of the tree dataframe}

\item{input_string}{OpenTree-compatible dataframe}

\item{treetype}{type of decision tree (decision or markov)}
}
\value{
evaluated model output
}
\description{
evaluate_model
\code{evaluate_model} evaluates the decision model
}
